Relocate ProviderFlags code and tests
