+++ cmd/internal/providers_test.go
@@ -3,6 +3,8 @@ package internal_test
 import (
 	"context"
 	"errors"
+	"flag"
+	"io"
 	"os"
 	"testing"
 
@@ -75,4 +77,36 @@ func TestCreateProviders(t *testing.T) {
 			}
 		}
 	})
+}
+
+func TestProviderFlags(t *testing.T) {
+	var gcp, aws, azure internal.StringSliceFlag
+
+	fs := flag.NewFlagSet("test", flag.ContinueOnError)
+	fs.SetOutput(io.Discard)
+	fs.Usage = func() {}
+
+	internal.ProviderFlags(fs, &gcp, &aws, &azure)
+
+	args := []string{
+		"-gcp.project=my-project",
+		"-azure.sub=my-subscription",
+		"-aws.profile=my-profile",
+	}
+
+	if err := fs.Parse(args); err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+
+	tests := map[*internal.StringSliceFlag]string{
+		&gcp:   "my-project",
+		&aws:   "my-profile",
+		&azure: "my-subscription",
+	}
+
+	for v, exp := range tests {
+		if len(*v) != 1 || (*v)[0] != exp {
+			t.Errorf("expecting one value (%q), got %v", exp, v)
+		}
+	}
 }
